requirements.txt
termcolor
opencv-python
moviepy
scikit-image
IPython
tornado
simplejson
torch==1.7.1
torchvision==0.8.2
tensorboard
